refactor: move storefront lib
