+++ starters/shopify-meilisearch/app/product/[slug]/draft/page.tsx
@@ -1,13 +1,13 @@
-import { unstable_cache } from "next/cache"
+import { Suspense } from "react"
 import { draftMode } from "next/headers"
 import { notFound } from "next/navigation"
-import { Suspense } from "react"
-import { storefrontClient } from "clients/storefrontClient"
 import type { CommerceProduct } from "types"
 
 import { Breadcrumbs } from "components/breadcrumbs"
 
+import { getAdminProduct, getProductByHandle } from "lib/shopify"
 import type { PlatformProduct } from "lib/shopify/types"
+
 import { getCombination, getOptionsFromUrl, hasValidOption, removeOptionsFromUrl } from "utils/product-options-utils"
 import { BackButton } from "views/product/back-button"
 import { SimilarProductsSection } from "views/product/similar-products-section"
@@ -95,13 +95,11 @@ async function ProductView({ slug }: { slug: string }) {
 async function getDraftAwareProduct(slug: string) {
   const draft = draftMode()
 
-  let product = await storefrontClient.getProductByHandle(removeOptionsFromUrl(slug))
+  let product = await getProductByHandle(removeOptionsFromUrl(slug))
   if (draft.isEnabled && product) product = await getAdminProduct(product?.id)
 
   return product
 }
-
-const getAdminProduct = unstable_cache(async (id: string) => storefrontClient.getAdminProduct(id), ["admin-product-by-handle"], { revalidate: 1 })
 
 function makeBreadcrumbs(product: PlatformProduct) {
   const lastCollection = product.collections?.findLast(Boolean)
