+++ starters/shopify-meilisearch/views/product/product-added-alert.tsx
@@ -12,7 +12,7 @@ import { type CurrencyType, mapCurrencyToSign } from "utils/map-currency-to-sign
 export function ProductAddedAlert({ className }: { className?: string }) {
   const router = useRouter()
   const { product, combination } = useAddProductStore()
-  const { openCart, preloadSheet, cart } = useCartStore()
+  const { openCart, preloadSheet, cart, checkoutReady } = useCartStore()
 
   if (!product || !combination) return null
 
@@ -44,7 +44,7 @@ export function ProductAddedAlert({ className }: { className?: string }) {
           <Button onMouseEnter={preloadSheet} onClick={openCart} variant="outline" className="bg-white transition-all hover:scale-105">
             View cart
           </Button>
-          <Button variant="default" onClick={() => router.push(cart?.checkoutUrl!)} className="rounded-md px-10 py-4 transition-all hover:scale-105">
+          <Button variant="default" onClick={() => router.push(cart?.checkoutUrl!)} className="rounded-md px-10 py-4 transition-all hover:scale-105" disabled={!checkoutReady}>
             Checkout
           </Button>
         </div>
